requirement.txt
mcp
Appium-Python-Client
selenium
requests